evaluation/pyproject.toml
[tool.poetry]
name = "evaluation"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
python-dotenv = "^1.2.1"
fastapi = "^0.120.3"
uvicorn = "^0.38.0"
chunking-evaluation = { git = "https://github.com/Chunkwise/chunking_evaluation_refactored.git" }
boto3 = "^1.40.69"
chunkwise-core = {path = "../core", develop = true}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

